xgettext: Address review comments for #243
