+++ i18n-helpers/src/xgettext.rs
@@ -14,6 +14,8 @@
 
 //! This file contains main logic used by the binary `mdbook-xgettext`.
 
+mod paths;
+
 use std::collections::HashMap;
 use std::io;
 use std::path::{Path, PathBuf};
@@ -27,8 +29,6 @@ use polib::catalog::Catalog;
 use polib::message::{Message, MessageMutView, MessageView};
 use polib::metadata::CatalogMetadata;
 use pulldown_cmark::{Event, Tag, TagEnd};
-
-const POT_EXTENSION: &str = "pot";
 
 /// Strip an optional link from a Markdown string.
 fn strip_link(text: &str) -> String {
@@ -183,10 +183,10 @@ where
     // `summary.pot` file within the default directory for chapters
     // without a corresponding part title.
     let mut unique_path_builder = UniquePathBuilder::new(depth);
-    unique_path_builder.maybe_push("summary");
-    unique_path_builder.maybe_push("summary");
+    unique_path_builder.push("summary");
+    unique_path_builder.push("summary");
     catalogs.insert(unique_path_builder.get(), catalog);
-    unique_path_builder.maybe_pop();
+    unique_path_builder.pop();
 
     // Next, we add the part-title and chapter contents.
     for item in &ctx.book.sections {
@@ -196,8 +196,8 @@ where
             // exists within.
 
             // Exit current part title and enter the new one.
-            unique_path_builder.maybe_pop();
-            unique_path_builder.maybe_push(title);
+            unique_path_builder.pop();
+            unique_path_builder.push(title);
         } else if let BookItem::Chapter(chapter) = item {
             // Add the contents for all of the sub-chapters within the
             // current chapter.
@@ -245,7 +245,7 @@ fn get_subcontent_for_chapter(
     unique_path_builder: &mut UniquePathBuilder,
 ) -> Vec<Chapter> {
     // Enter current chapter.
-    unique_path_builder.maybe_push(&chapter.name);
+    unique_path_builder.push(&chapter.name);
 
     let chapter_info = chapter.path.as_ref().map(|path| Chapter {
         content: chapter.content.clone(),
@@ -266,151 +266,8 @@ fn get_subcontent_for_chapter(
         .collect();
 
     // Exit chapter.
-    unique_path_builder.maybe_pop();
+    unique_path_builder.pop();
     sub_chapter_info
-}
-
-pub mod paths {
-    use super::POT_EXTENSION;
-    use std::collections::HashMap;
-    use std::path::PathBuf;
-
-    pub struct UniquePathBuilder {
-        current_path: PathBuf,
-        current_depth: usize,
-        max_depth: usize,
-        frequency_map: HashMap<PathBuf, usize>,
-    }
-
-    /// Mimics the `push` and `pop` functionalities of `std::path::PathBuf`
-    /// with, respectively, the [`maybe_push`] and [`maybe_pop`] methods,
-    /// so that:
-    ///
-    /// 1. The path is never extended past `max_depth` depth.
-    /// 2. Whenever [`maybe_push`] is invoked, the base directory/file of
-    ///    the path is potentially altered to make the path unique among
-    ///    the ones produced thus far.
-    /// 3. All strings passed to [`maybe_push`] are normalized by retaining
-    ///    only non-alphabetic characters and hyphens.
-    ///
-    /// The disambiguation is deterministic (i.e., the same sequence of
-    /// operation is guaranteed to always produce the same paths) and not
-    /// based on randomization, therefore its stability doesn't depend on
-    /// the stability of external libraries.
-    ///
-    /// The [`get`] method returns the current path with a .pot extension.
-    ///
-    /// The spacial case of `depth` equal to 0 is treated differently by
-    /// returning the `DEFAULT_DEPTH_0`.pot filename.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use mdbook_i18n_helpers::xgettext::paths::UniquePathBuilder;
-    ///
-    ///
-    /// let mut unique_path_builder = UniquePathBuilder::new(2);
-    ///
-    /// fn get_str(pb: &UniquePathBuilder) -> String {
-    ///     pb.get().to_str().unwrap().into()
-    /// }
-    ///
-    /// unique_path_builder.maybe_push("foo"); // Pushes "foo"
-    /// unique_path_builder.maybe_push("bar"); // Pushes "bar"
-    /// assert_eq!(get_str(&unique_path_builder), "foo/bar.pot");
-    ///
-    /// unique_path_builder.maybe_push("baz"); // No-op for the user
-    /// assert_eq!(get_str(&unique_path_builder), "foo/bar.pot");
-    ///
-    /// unique_path_builder.maybe_pop(); // No-op
-    /// unique_path_builder.maybe_pop(); // Pops "bar"
-    /// unique_path_builder.maybe_pop(); // Pops "foo"
-    /// unique_path_builder.maybe_push("foo"); // Pushes "foo_1"
-    /// unique_path_builder.maybe_push("bar"); // Pushes "bar"
-    /// unique_path_builder.maybe_pop(); // Pops "bar"
-    /// unique_path_builder.maybe_push("bar"); // Pushes "bar_1"
-    /// assert_eq!(get_str(&unique_path_builder), "foo_1/bar_1.pot");
-    /// ```
-    impl UniquePathBuilder {
-        const DEFAULT_DEPTH_0: &str = "messages";
-
-        pub fn new(max_depth: usize) -> Self {
-            Self {
-                current_path: PathBuf::new(),
-                current_depth: 0,
-                max_depth,
-                frequency_map: HashMap::new(),
-            }
-        }
-
-        fn format_path_with_counter(&self) -> PathBuf {
-            match self.frequency_map.get(&self.current_path) {
-                None | Some(&0) => self.current_path.clone(),
-                Some(&cnt) => {
-                    let mut file_name = self.current_path.file_name().unwrap().to_os_string();
-                    file_name.push(format!("_{}", cnt));
-                    self.current_path.with_file_name(file_name)
-                }
-            }
-        }
-
-        pub fn get(&self) -> PathBuf {
-            if self.max_depth == 0 || self.current_depth == 0 {
-                PathBuf::from(Self::DEFAULT_DEPTH_0).with_extension(POT_EXTENSION)
-            } else {
-                self.format_path_with_counter()
-                    .with_extension(POT_EXTENSION)
-            }
-        }
-
-        pub fn maybe_push<T: AsRef<str>>(&mut self, s: T) {
-            self.current_depth += 1;
-            if self.current_depth > self.max_depth {
-                return; // Exceeded max depth, do nothing.
-            }
-            // Extend current path with new component.
-            self.current_path = self.format_path_with_counter();
-            self.current_path.push(slug(s.as_ref()));
-            // Update frequency of new path.
-            self.frequency_map
-                .entry(self.current_path.clone())
-                .and_modify(|cnt| *cnt += 1)
-                .or_insert(0);
-        }
-
-        pub fn maybe_pop(&mut self) {
-            if self.current_depth == 0 {
-                return; // Maybe panic instead?
-            }
-            if self.current_depth <= self.max_depth {
-                self.current_path.pop();
-            }
-            self.current_depth -= 1;
-        }
-    }
-
-    // Trim a string slice to only contain alphanumeric characters and
-    // dashes.
-    // If the resulting string is empty, returns "null" instead.
-    fn slug(title: &str) -> String {
-        // Specially handle "C++" to format it as "cpp" instead of "c".
-        let title = title.to_lowercase().replace("c++", "cpp");
-        let title = title
-            .split_whitespace()
-            .map(|word| {
-                word.chars()
-                    .filter(|&ch| ch == '-' || ch.is_ascii_alphanumeric())
-                    .collect::<String>()
-            })
-            .filter(|word| !word.is_empty())
-            .collect::<Vec<_>>()
-            .join("-");
-        if title.is_empty() {
-            String::from("null")
-        } else {
-            title
-        }
-    }
 }
 
 #[cfg(test)]
@@ -1257,8 +1114,8 @@ mod tests {
                 PathBuf::from("foo.pot"),
                 vec!["src/foo.md:1", "src/bar.md:1"],
             ),
-            (PathBuf::from("foo_1.pot"), vec!["src/foo-back.md:1"]),
-            (PathBuf::from("foo_2.pot"), vec!["src/oh-no-foo-again.md:1"]),
+            (PathBuf::from("foo-1.pot"), vec!["src/foo-back.md:1"]),
+            (PathBuf::from("foo-2.pot"), vec!["src/oh-no-foo-again.md:1"]),
         ]);
 
         assert_eq!(expected_message_tuples.keys().len(), catalogs.len());
